Addresses: https://github.com/sa-lee/plyranges/pull/80#discussion_r426367218
**Solution:**

`expand_hits` now has an additional optional argument
`hits_mcols_to_keep` (default: `NULL`) which takes a named vector where names()
corresponds to the destination name, and the values correspond to the column
names in `hits`. This produces an error if a destination column already exists
with the same target name (addresses suffix issue for joins).

**Remaining Issues:**

This feature does not work when `return_data_frame ==
TRUE`, since I'm not familiar enough with the `plyranges` codebase, I am not
sure how this argument is really used, and I didn't look around too hard. I
_think_ you could just move the block down after the `hits_mcols_to_keep` logic
to include the metadata without issues, but I didn't want to make that change
and break something else since this is deep in the internals.